PsychEyelinkDispatchCallback: Fix target drawing for large target dot sizes.
Code unconditionally used ‘DrawDots’, even for dots which might be
beyond the limit of the GPU.
Now use ‘DrawDots’ for dots smaller than 10 pixels diameter, ‘FillOval’
for others.
The limit is hardcoded to a safe 10 pixels setting that should work on
Intel, Nvidia and AMD.
Query of the GPU limit would be possible but a bit more inefficient -
or more work to implement efficiently.