Screen('TextFont'): Allow simultaneous setting of font name and font style.
Do ‘TextFont’ and ‘TextStyle’ in one atomic call. This is needed to
work around broken fonts on OSX in combination with OSX broken font
handling. See forum msg #17286 for the problem.
If a font does not support standard textStyles, it wasn’t possible on
OSX to select it by name, because name selection and style selection
were different calls and each call failed font validation individually,
because it would have had to transition through an invalid state in the
[fontFamilyName, fontStyle] space. Setting both atomically allows to
jump from one valid state to the next valid state and validation
succeeds. This is not needed on Linux or Windows, but now implemented
for compatibility.
On old version of Screen() selection of such a font by fontNumber was
still possible as a workaround.

-> Screen for OSX Matlab & Octave rebuilt and tested.