PsychScriptingGlue: Fix global state bleeding between mex files.
Make prhs/plhsGLUE et al. static so their state can't bleed from
one PTB mex file to the other. This was the case when Screen()
was calling DataPixx() from within its post-swap hooks to do
timestamping with DataPixx. Datapixx was globbering the state
of Screen. This was the reason for the errors in VBLSyncTest.m
when Datapixx + Octave was used. The error didn't happen on Matlab
because there we compile mex files to only export mexFunction()
but nothing else into the process symbol space.

-> Fixed.
-> Screen for 64-Bit Octave on Linux rebuilt and tested.
-> VBLSyncTest: Remove all workarounds.