Screen/Linux: Add workaround for broken pageflip events on nouveau in 3.13+
nouveau-kms on Linux 3.13 - 3.14 - and possibly future 3.15 returns
faulty data in kms-pageflip completion events. Underlying bug is
diagnosed and first patches exist to fix it, but they didn't make the
deadline for Linux 3.14. The bug was introduced in Linux 3.13, so
the 3.13 kernel of Ubuntu 14.04 LTS will be affected.

Add workaround for OML_sync_control timestamping: If nouveau-kms on
3.13 - 3.15 is detected, we still use glXWaitForSbcOML to wait for
doublebuffer swap completion, but then use glXGetSyncValuesOML to
get current vblank count and timestamp as possible replacements for
the broken/invalid msc and ust from the pageflip event. We validate
the values and if they are safe to use we use them. Otherwise we
trigger fallback to classic mmio beamposition timestamping which
is a tad more noisy and adds extra execution overhead, but safe to
use.

-> Screen for 64-Bit Octave on Linux rebuilt and tested.