Screen/OSX: Always use Cocoa+NSOpenGL for fullscreen windows on OSX 10.8+
This is an implicit application of the kPsychUseAGLForFullscreenWindows
setting when running on OSX 10.8 or later.

Testing on OSX 10.8 with dual-display stereomode 10 showed that
as soon as a 2nd onscreen fullscreen window is opened on a second
display, OSX 10.8+ will do the most brain-damaged thing i've seen
in a while: It will enable page-flipping for bufferswaps on the
second window, but *disable* page-flipping on the first window on
which page-flipping was so far enabled. Iow., the first opened
window is now subject to desktop composition with potentially
horrible effects on timing and timestamping precision and reliability!!
Also 10 bpc framebuffer mode won't work anymore because the compositor
screws with the alpha-channel of onscreen windows, stripping it
effectively away! This is how we actually found this nugget of a
hilarious bug, and it took us only a week.

Presence of the bug - or its absence after this workaround - can
be verified with OSXCompositorIdiocyTest().

What this fix probably doesn't fix:

- The black screen at end of a session.
- The completely messed up swap handling of NVidia gpu's under 10.9

It is also conceivable that the fix corrupts timing in a different
way, but i can't test this, as a) i don't have OSX 10.8 and b)
My 10.9 has the totally f***d NVidia drivers. However, some
testing by Nicolas on AMD cards suggests this fixes the problems
on at least 10.8.