Screen/OSX: Disable VBLIrq/CoreVideo timestamping fallback by default.
Use VBLTimestampingmode 0 by default for the 3.0.12+ series, instead of
mode 1.
This disables use of VBL-IRQ/Shmem fallback on OSX 10.6 Snow Leopard,
as we can’t test it for proper functioning anymore. Might still work
ok, but who knows?
This disables use of Apple’s broken CoreVideo/CVDisplayLink
timestamping on 10.7+. It was unreliable at best, but also causes
random crashes on some setups. This is not fixable by us and i’m sick
of seeing bug reports for it, so disable by default.

Also change recommendation messages to only recommend installation of
the PsychtoolboxKernelDriver for proper timestamping if we are not
running on a Intel GPU, where the kernel driver is disabled for
security reasons.

-> Screen for 64-Bit Octave on OSX rebuilt and tested with
PTKD/non-PTBKD, Nvidia vs. Intel.
-> help texts updated.