Core: Rename PsychScriptingGlue.cc -> .c for pure C build.
Changing PsychScriptingGlue's suffix from .cc to .c will
treat it as a C language file instead of a C++ language
file and thereby compile it as C instead of C++.

In the past, over 5 years ago, we needed this file to be
compiled as C++ for Octave compatibility, but we don't
need this any longer. A C build will do just fine. A C
build also avoids occassional trouble with C++ compilers
and C++ header files. C++ build was a recurring source
of build trouble with new compiler/framework/octave versions.

-> Renamed file.
-> Updated all make files and cleaned them up a bit in the process.
-> Test rebuild many mex files for many platforms.