Screen/Linux: Refine nouveau-kms broken pageflip event workaround.
Use video refresh interval * 0.8 if refresh interval is known post
refresh calibration. Use 7 msecs as in the past if refresh interval
not yet known, ie., during calibration itself. This is more adaptive
to true refresh rate.

Also blacklist all kernels >= 3.13 for now. We know 3.13 and 3.14
are definitely broken on nouveau-kms, but we don't know yet when
the fix will land - 3.15? 3.16? Later? Better play it safe and
assume all future kernels are broken until we know in which kernel
release the bug was finally fixed.