Screen: Silence various compiler warnings.
The compiler found one real bug in Screen('GetImage'), although
one that was likely benign and never triggered.

There were various false positives about use of uninitialized
variables, mostly due to PsychErrorExitMsg() not being recognized
as the equivalent of a return() statement. This is just too convoluted
for the compiler to prove. Added some dummy assignments to silence
this noise and similar stuff that the compiler can't know is right.