Screen/GStreamer: Only allow output of audio output device on Linux.
Turns out this is not supported on Windows at all and only in a
completely meaningless and source-code incompatible way on OSX, so this
is only useful for Linux + GStreamer.

-> Screen and IOPort for 64-Bit Octave/Matlab on OSX rebuilt and tested.
-> Setting audio sink on OSX works ok’ish, although the osxaudiosink
doesn’t respect volume settings and the device setting is essentially
useless on my MacbookPro with 10.9.2, as the only valid value is
device=0, which is the default anyway. Any other value causes failure.

-> In other news, GStreamer movie bus message delivery works on OSX
with Octave 3.8.0 + GUI, so either the problem is restricted to Linux
or to the specific set/versions of GStreamer plugins on my Ubuntu
14.04-LTS machine.