Screen: Disable RTLD_GLOBAL export of Screen's symbols on Linux.
Disable the dlopen() self-reloading() of Screen. It had the
purpose of making Screen's symbols globally available to other
shared libraries. However, the reason for doing this is long
gone (ListenChar magic sauce for Octave + GUI) as it proved
too fragile.

Now the code just causes trouble because mex files like
moglcore.mex would use Screen's symbols (e.g., for GLEW stuff)
and keep Screen.mex loaded into the process when it should
unload on 'clear Screen'. We disable this useless code, but
leave it in place to document the approach for future more
targeted use.

-> This prepares the latest workaround in the next commit
   for an annoying Mesa bug.

-> Remove some orphaned prototype from ScreenCloseAll().