PsychRTBox.m: Fix for firmware version 5.0
v5.0 firmware doesn't seem to send back acknowledge char 'e' anymore
for event 'e'nable. So don't expect an ack for 'e' on version 5.

Is this a complete fix? I don't know, because that f$#%# firmware
and underlying protocol gets changed over and over to make it a
bit worse with each iteration.