Screen/Linux: Fix off-by-one error in window size query for kPsychGUIWindow's
PsychOSProcessEvents() recalculates window size from queried results on
kPsychGUIWindows to account for live resizing of the GUI window by the
user. On Linux it erroneously subtracted 1 from width and height, when
width and height was already correct, leading to some failure in some
tests, because all code suddenly assumed a 1 pixel smaller window.
Fixed in this commit.