DrawFormattedText: Add new sx options 'wrapat' and 'justifytomax'
These options try to justify the width of text to both left- and right
border, so width of text appears roughly even. ‘wrapat’ justifies to a
fixed line width of ‘wrapat’ characters, whereas ‘justifytomax’
justifies to the width of the widest line of text in a given multi-line
text string.

This is achieved by changing the width of blanks in each line of text
to “rescale” the width of the line of text. If lines deviate more than
global ptb_drawformattedtext_padthresh
from the reference width, they are not stretched, to avoid ridiculous
looking results.
Default threshold is 33%.

This works ok’ish if the text is already separated via newline
characters into lines of roughly equal length.

DrawFormattedTextDemo updated to demonstrate this.