help update for SyncTrouble and PsychImaging.
Mention in help SyncTrouble how to enable unredirect fullscreen windows
under KDE/KWin, now that we are back to old-style override redirect for
KDE multi-display setups.
Mention in PsychImaging that radeon.deep_color=1 must be passed to the
Linux 3.16 kernel command line to enable deep colour support, as it is
disabled by default.