DrawFormattedText: Small adjustment for Linux text renderer.
Must use X as proxy for a blank when calculating the bounding
box of an isolated blank, as the Linux renderer optimizes the
blank away and we end with an empty bbox and therefore wrong
base-advance and compressed text.

This works because justification is only workable with monospaced
fonts anyway, where an X has the some horizontal width as a ' '
blank.

(cherry picked from commit 80f9a798ca0940097877de4d9c9fbed6457a3c3a)