Fix up ComputeRMSE the way I should have to being with.
PTB contained a routine called ComputeRMSE that did not actually compute the RMSE (it computed a root fractional sum of squared errors.)

At one point I put in a warning message to this effect, but that was fairly annoying to see all the time.  So

I wrote a ComputeFSSE() that does what ComputeRMSE used to.
I changed all calls to ComputeRMSE in the toolbox to calls to ComputeFSEE().
I changed the warning in ComputeRMSE to suggest that the user change the call to one to ComputeFSSE().

In doing all this, I discovered that PTB contained two routines ComputeRMSE, one in PsychGamma and one in
PsychSignal (the latter was shadowed and never called).  I removed the one in PsychSignal (which in fact
computed the same thing as the one in PsychGamma.)

I also updated the relevant Contents.m files.

DHB

Signed-off-by: David Brainard <brainard@psych.upenn.edu>