Screen/Linux: Add safeguard against negative msc values.
On two occassions glXWaitForSbcOML() returned a negative msc
value, which is legal, as msc is a signed 64 Bit integer, but
our remaining code can't deal with this, so simply strip the
sign bit from returned msc's, so they are always positive and
incrementing from the perspective of our code, so we are
protected against such a "underflow by overflow" scenario.

It was not clear in those two isolated cases why suddenly
negative (but incrementing) values were returned, as values
large enough to overflow are very unlikely in practice. A
restart of the X-Server via logout/login resolved the problem.
Maybe somewhere a zero initialization is missing in XOrg 1.15?
That would technically be not a bug, as the "start point" is not
defined and could be in the high value range, but it would still
be confusing to clients.