Screen: Support new imagingMode flag kPsychNeedOtherStreamInput.
When the flag is provided and a stereomode is active, this will
bind the complementary input image to texture unit 1 during
processing of a stereo processing chain. For left-view processing,
the input image of the right view channel is exposed, during right-
view processing, the input image of the left view channel is exposed.

This allows for things like stereo crosstalk reduction shaders
which need access to the "other eyes" input image to use it for
their suppression strategy.