Screen/Linux: Add final workaround setup for nouveau-kms pageflip bugs.
Rework this again. The timestamping workaround for nouveau-kms pageflip
completion event bugs in Linux 3.13-3.15 can be enabled via the
conserveVRAM setting.It will also get auto-enabled if a broken kernel
is detected, ie., a kernel with a version prior to the versions we know
to contain the fixes. This is easily done for 3.14 and 3.15. Linux 3.13
needs special treatment, depending if it is a longterm stable kernel, or
a Ubuntu distribution kernel based on a longterm 3.13 stable kernel.

Now we have Ubuntu specific detection logic in place to detect if we
are on a vanilla 3.13 kernel or Ubuntu one. And specific detection logic
to decide if the workaround is needed or not.

The current situation mid August 2014 is that the bug is fixed in all
upstream/mainline/vanilla stable kernels (3.13.11.5+, 3.14.12+, 3.15.5+),
and in all Linux 3.16+ kernels anyway since initial release.

The Ubuntu 14.04-LTS distribution kernel still isn't fixed, one month
after Ubuntu's kernel team fixed the 3.13.11.5+ stable longterm kernel
on which the 14.04-LTS kernels are based. This takes longer than i
expected, so we add this Ubuntu special case, which can automatically
decide about the status of the 14.04-LTS kernel and act appropriately.

Only source code updated, to allow cherry-picking into ptb 3.0.11.