Add prototype implementation of stereo crosstalk reduction.
This adds a new, but so far hidden/undocumented PsychImaging
task 'StereoCrosstalkReduction', and associated demo code inside
ImagingStereoDemo.m.

Requesting ...

 PsychImaging('AddTask', 'LeftView', 'StereoCrosstalkReduction', reduceCrossTalk);
 PsychImaging('AddTask', 'RightView', 'StereoCrosstalkReduction', reduceCrossTalk);

...would load the new StereoCrosstalkReductionShader.frag.txt GLSL shader, set it
up with parameter 'reduceCrossTalk' (whatever that is) and attach it to a properly
configured imaging pipeline. It's purpose in life would be to somehow reduce
crosstalk in stereo images.

Currently the shader(s) get attached at the top of the per-view image processing
chains, one shader for left-eye, and one for right-eye. This way the input images
to the shaders are the (panel scaled/fitted and multisample anti-aliased) stimulus
images as drawn by usercode. We use one individual shader per eye, so each shader,
although containing the same code, can be parameterized differently for the two
eyes.

The shader must be a single-pass shader, as the content of the 2nd stereo channel
(the one whose ghost images shall be suppressed) can't be carried through multiple
pipeline slots or passes.

Also shader placement isn't yet optimal in case other operations like image flipping,
geometric undistortion etc. would be also part of the image processing chains. This
may need more work, once the basics are proven to be working, but as a starter it
should be good enough.

The shader itself is currently just a "here starts your task" implementation,
which doesn't do something meaningful, apart from demonstrating to Diederick how
to access the two input images and create an output image. Filling the shader with
some meaning is Diedericks task.

This commit also contains a properly compiled Screen mex file for 64-Bit Octave on
Linux to make this work.

Recompile for Screen mex files for other runtimes and os'es is pending for a
separate commit.