Screen/OSX: Remove old shmem vbl-irq timestamping for OSX 10.6
This only worked up to and including OSX 10.6 Snow Leopard.
We don’t care about (or test on) 10.6 anymore, so let this die.
Now the only opt-in fallback for lack of functional beamposition
timestamping on OSX is the buggy and unreliable CoreVideo timestamping.
This is an opt-in, it is off by default and must be enabled via
Screen(‘Preference’,’VBLTimestampingmode’, 1) by adventurous or
desperate users if they really want it.