Screen/OSX: Remove native CGDisplayBeamposition() queries.
Support for this function was sicklish on 10.7, horribly broken on 10.8
and is non-existent on 10.9 going forward. Thanks Apple!
On 10.6 they work, but we don’t care about that OS anymore.

-> Let it r.i.p.
-> Now the PsychtoolboxKernelDriver is required for any kind of high
precision timestamping, except for use on Intel GPU’s where no such
thing as high precision timestamping exists anymore, as use of the
kernel driver there is dangerous for system stability and health.