Screen: Also ask OS to provide 16 bpc framebuffers at pixelSize 48.
First try to get 16 bpc / 64 bpp from OS, then fall back to our
own method on Linux.

At the moment no OS supports this natively, so we will effectively
fail on Windows, OSX, and Linux with non-AMD gpu's, and trigger our
fallback on Linux + X11 + AMD gpu's if mmio access is enabled.

But it's good to prepare for the future, so implement this already.

-> Screen for 64-Bit Octave on Linux rebuilt and lightly tested.
-> Some whitespace cleanup in source files i touched.