moglcore: Add updated glext.h, derived glext_edit.h and glu_edit.h
The glu.h coming from OSX is not parseable by our code generator
anymore, as Apple decided to sprinke DEPRECATED macros all over
the place to declare their intention to make use of GLU in the
future as painful as possible. Therefore we include our own
untainted copy of it, as permitted by its license.

glext.h is up to date as of OpenGL version 4.5 and extensions.
glext_edit.h removes some useless defines etc. to make it better
parseable.

This glext_edit.h has functions which are not supportable by
current GLEW 1.11 at build time, so this is just a baseline
for further stripping off of functions.