Screen: Make max allowable vtotal configurable.
Instead of hard-coding that vtotal < vactive * 1.25, allow
to replace 1.25 by a configurable Screen('Preference') setting.

Default is 1.25 as in the past, but Screen('Preference','VBLEndlineOverride')
now allows to optionally specify a different multiplier to tweak
this limit.

This is because 1.25 is a reasonable limit for detecting broken
beamposition queries, but it is too low for some special video
modes used by VPixx Inc's ViewPixx-3D panel, where the vblank
can be up to 50% of vactive, ie., vtotal can be up to 1.5x
vactive. This is part of the bag of tricks used to give shutter
goggles enough switching time for frame sequential stereo.

Once we implement proper logic in PsychImaging/PsychDatapixx to
detect this use case, this should fix false failures of the
beamposition query correctness checks with such devices.

See forum message 18143 for description of the problem.

-> Screen for 64-Bit Octave and 64-Bit Matlab on Linux rebuilt.
-> Other platforms pending.
-> Also quite a bit of whitespace cleanup in preference glue.
-> Also ineffective mmap() read-only fallback for intel gpu's on
   Linux. Thought this would help against EAGAIN mapping failures,
   but doesn't. Anyway, not a real loss, given we have OML_sync_control
   on Intel.