PsychImaging: Allow a 50% vblank height on VPixx devices in stereo mode.
VPixx devices in frame sequential stereo mode may use vblank intervals
up to 50% the height of vactive. This would normally cause a detection
of broken beampos queries, as we only allow up to 25% by default.

Raise the limit for max valid vblendline to 50% or up to 1.5 * vactive
if a VPixx device in frame sequential stereo mode is in use.