Screen: Add new screen glue function PsychGetScreenPixelSize()
Reports size of screens frame buffer in pixels, instead of logical size
units (points), as PsychGetScreenSize() / GetScreenRect() /
GetScreenGlobalRect() do.

In the past, size in points was equal to size in pixels, so no need for
this.
On OS X with Retina displays (HiDPI displays) the logical/virtual size
in points can be different from the “physical” size in pixels. E.g., on
a Retina panel in “best for Retina” scaled mode, 1 point == 2 pixels.
Sometimes we need the true size of the display in pixels, therefore
this new function.
On Linux and Windows, the new function is just an alias for
PsychGetScreenSize().