Screen/OSX: Track windowRecord->globalRect and ->rect in glue code.
Cocoa glue now not only updates globalRect with virtual global display
coordinates in resolution independent unit of points, but also updates
the true backbuffer size of the window in pixels. Assignment is done at
OpenGL context assignment and for GUI windows at each move/resize.

This is needed because ‘rect’ must contain the true size of the windows
backbuffer in pixels, and this size can differ from the size in points
on HiDPI/Retina displays.