Screen('Rect'/'WindowSize'): Honor realFBSize flag also for screens.
In the past the flag was only taken into account for onscreen window
handles.
Now it also works for display screens, aka screenIds. This is important
so M-Files like PsychImaging() can query the true size of the frame
buffer in pixels (as opposed to points) before an onscreen window is
opened. Needed for proper PanelFitter setup and similar tasks on Retina
displays.