Screen/OSX: Retina mode: Unconditionally request a frame buffer of native resolution.
On OS X we always unconditionally request an OpenGL window backbuffer
the full true scanout size of the video mode in pixels. This makes sure
that if a Retina panel is connected, and a scaled video mode (HiDPI) is
in use, our OpenGL frame buffer for the onscreen window has the full
native panel resolution.

-> Allows to take full advantage of HiDPI Retina displays.
-> Should avoid any rescaling by the OS and thereby make visual onset
timing and timestamping less worse - i’d say good, but that seems to be
an elusive goal on the most broken OS in existence. At least it
shouldn’t be worse than on a non-Retina panel.

Remaining issues:

* If GUI windows get manually resized by user interaction, we get lots
of pixel trash, although all our window and rendering parameters get
properly updated. Possibly a bug in Apples Retina implementation when
the compositor is active? Nothing we could do about it, except maybe
disable Retina rendering for GUI windows?

* Mismatch between reported mouse cursor locations and rendering
locations, as the cursor position seems to be reported in virtual
points, not true pixels.

* Should we use our panel fitter automatically by default on  a Retina
panel? Otherwise resolution and rendering load will be very high on
such panels, which causes tiny rendering on old Psychtoolbox scripts
and reduced performance due to the possibly overtaxed gpu.

-> Screen for 64-Bit Octave/Matlab on OS X rebuilt and tested.