Screen: Default to scaled low-res mode on Retina displays.
If a Retina style HiDPI display with a scaled video mode is detected,
enable the panel fitter by default to provide backward compatible
rendering and display behaviour. The apparent resolution of the frame
buffer is artificially limited to the resolution of the scaled video
mode (typically half horizontal and vertical resolution for about 1
quarter the total pixel resolution), then the panel fitter upscales the
stimulus image to full native Retina resolution before display. This
saves computation time on the GPU and some coordinate system conversion
headaches for user code, but obviously also throws away the high
resolution advantage.
The behaviour is basically the same as with PTB 3.0.11 and how OS X
behaves by default for any application which is not specifically
prepared for Retina support.

A new PsychImaging task ‘UseRetinaResolution’ allows to override this
and get the full native panel resolution, performance impact and
coordinate transformation fun included.

If the panel fitter is used by user code then this auto-scaling won’t
kick in and user code is fully responsible for setting up a scaling
that suits it.

-> Code is generic for all operating systems, but currently only
affects OS X with Retina displays.
-> Screen for OS X 64-Bit Octave/Matlab rebuilt and lightly tested.