RemapMouse(): Handle Retina display mode remapping.
Need to convert GetMouse() reported positions in points to equivalent
positions in pixels, before we go through the other inversion and panel
fitter remapping.

-> This automatically no-ops on non-Retina display modes or Retina
compat modes, because either the backingScaleFactor is 1 or it cancels
itself out with the inversion of the panel fitter transformations.

-> Need it for use of geometric image transformations as applied via
PsychImaging, and for regular panel fitter operations, and for using
Retina displays at full resolution.
-> Don’t need it for regular operation with regular display modes or in
Retina backward compatibility mode, just as in PTB 3.0.11.