ImagingStereoDemo, ImagingVideoCaptureDemo: Fix quirks uncovered by Retina stuff.
ImagingStereoDemo: Restricting the processing ROI is harmful in yet
another use case - Retina stuff. This did more harm than good in that
demo, so comment it out.
ImagingVideoCaptureDemo: Coordinate transform for blur region was wrong
or at least icky, but multiple “bugs” cancelled each other out. Not so
with Retina anymore, so fix the transformation to work in all cases.