Screen/Linux: Always override-redirect fullscreen windows on KDE multi-display.
As i learned from Martin Graesslin, KWin's maintainer at XDC 2014, KWin deliberately
doesn't unredirect fullscreen windows by default on Intel gpus, to work around bugs
in some versions of the Intel-ddx, that's why the Intel trouble. Therefore we must use
new-style for KDE to unredirect fullscreen windows to enforce compositor disable via
our XAtoms. However, this doesn't work for multi-display spanning windows due to what
i think is a KWin bug. On a dual-display setup, the right half of the window on the
2nd display gets cut off - the desktop is showing instead. No hack in the code or any
KWin rule i tried could prevent this cutoff, although the window itself is reported to
PTB as spanning both desktops - that's why this is probably not intentional KWin behaviour,
but a bug. To work around this we enforce old-style setup again on KWin + Intel + fullscreen
if there are at least two active video outputs connected to our target x-screen, aka
we are on a multi-display setup. This means the user has to manually disable desktop
composition if dual/multi-display stimulation on a Intel gpu under KDE/KWin is needed
with good presentation timing, as the manual "user says so" override is the only method
that worked.

-> Screen for 64-Bit Octave on Linux rebuilt and tested under KDE+Intel gpu, single
   and multi-display.

-> Single display is fully automatic as in the past.
-> Dual display needs manual compositor disable via control panel settings
   or SHIFT+ALT+F12 toggle by user.
-> Other gpu's than Intel are unaffected and KWin's unredirection of fullscreen
   windows works.