Screen/Linux: Be more lenient in detecting INTEL_swap_events.
Turns out that at least under nouveau-ddx with its glamor backend
enabled under DRI3/Present, the INTEL_swap_events extension is
fully supported and functional, but doesn't get advertised under
either the glXQueryExtensions() string, nor the server string. It
is only present in the client string. If this is spec conformant
or an omission on nouveau's side is unclear.

Our old matching code wouldn't enable use of this useful extension
under nouveau+glamor+dri3/present for this reason. Change the code
so it does. Now we try enabling the extension if it is advertised
in any of the GLX extension strings. Our code can easily handle
failure to enable the extension for real with no downsides, so this
should be all good.

-> Screen mex for 64-Bit Octave on Linux rebuilt and successfully
   tested under nouveau-ddx with dri3/present + glamor acceleration.