Add Weston .h/.c files or preliminary presentation_feedback client interface.
These are needed to use the presentation_feedback extension
for testing and development. Probably should be installed
somewhere properly instead being included into our source tree,
but for now, as this is not yet part of any official Wayland/Weston
distro and very much in flow, put it into our tree.

Licenses are BSD/MIT style, so they allow this.

-> Allows to compile our own client backend code.