Screen/Linux: Add support for Wayland windows that are transparent to user input.
Allow PsychDebugWindowConfiguration() to work better, by allowing user
input via mouse and touch to "fall through" the window, hitting windows
below the window.

This doesn't quite have the perfect result compatible with X11/OSX/Windows
as i hoped, because clicking on the window sends the mouse click to the
octave terminal window which gets raised on top of our onscreen window
due to "click to raise" mouse click handling. The transparency doesn't
seem to work for keyboard input, so we are not quite there, but closer...

-> Screen rebuilt and tested.