Screen/Linux: Implement video mode switching for Wayland backend.
We switch the mode "virtually" by patching up our mode list
to mark the mode we wish we had to be the mode we have ->
moving the "current mode" flag from the real mode to the
target mode -- after validation that is.

Then on a fullscreen window switch, our window picks up
the size and refresh rate of the "current mode" aka our
target mode as its own size and target rate and the
wl_shell_set_fullscreen() method hopefully does a mode
switch to satisfy our windows size and the requested
target refresh rate.

So far the theory. In practice resolution switching indeed
works nicely on the Weston drm backend, but the refresh
rate remains unchanged, although we select a mode that
was provided by Weston to us as a valid target.

Anyway, as good as it gets so far.

-> Screen for 64-Bit Octave rebuilt and tested on Weston/DRM
   on a dual-display setup.