Screen/Linux: Add global window transparency support for Wayland.
To implement PsychDebugWindowConfiguration, we need to be able
to apply a global alpha to the windows backbuffer, ie., post-
multiply per pixel alpha values with the global alpha derived
from the windowShieldingLevel set by PsychDebugWindowConfiguration.

Wayland 1.6 currently doesn't have protocol support for a global
alpha, so we use a post-processing pass in the imaging pipeline:

Add a new builtin function for alpha post-multiply to imaging
pipeline and call it from the finalizer chains as very last
post-processing step before OpenGL doublebuffer swap.

-> Screen for 64-Bit Octave on Linux rebuilt and tested on Wayland.
-> Also a little bit of indentation cleanup and tweak to some
   Wayland debug output verbosity level.