Add hardware OpenGL setup code for Matlab R2014b+ on Linux.
Matlab R2014b and later on Linux will detect if the systems
Mesa OpenGL library would be used. It doesn't like that for
its new OpenGL based plotting system because probably Mathworks
doesn't want to do the needed QA. Therefore Matlab switches to
its own - horribly outdated - Mesa X11 v7.2 software renderer.

Of course we don't like that as that thing is completely
unuseable for any vision science testing. Therefore the
PsychLinuxConfiguration.m setup script will now detect if
it is running on top of R2014b+ with the software renderer.
If so it will change Matlab's startup default of renderer to
use the system installed (Mesa) OpenGL library to make things
good again, and then prompt the user to restart Matlab.

Screen() now detects during OpenWindow if it is running on
this software 'Mesa X11' renderer and gives ample warning to
the user on how to fix the problem. It also aborts unless the
user forces acceptance of the software renderer via proper
'ConserveVRAM' flags 64 == kPsychUseSoftwareOpenGL.

-> Screen updated.
-> Docu and M-files updated, also some white-space cleanup.