Screen: In dc1394 capture, call tracker plugin after colorspace conversion.
Call the optional tracking plugin after performing an optional colorspace
conversion (YUV -> RGB or Raw-Bayer -> RGB), instead of before. This way
the plugin can be applied to reconstructed RGB data if such conversion
is needed. This simplifies the plugin a lot.

Screen for 64-Bit Octave on Linux rebuilt and tested.