Screen/GStreamer: Add hint to use recordingflags 8 to resolve issues.
Some video capture devices don't like the ready->paused->playing->ready
transition for video caps probing in 'OpenVideoCapture' and fail if a
similar sequence is later executed in 'StartVideoCapture', although they
worked in 'OpenVideocapture'. One shot only for this.

recordingflags & 8 skips that first sequence in 'OpenVideocapture',
thereby makes the second sequence in 'StartVideoCapture' work. Ergo,
if we detect failure in 'StartVideocapture' for paused->playing
transition and recordingflags wasn't set to 8, print out some hint
to the user to retry with a setting of 8.

Resolves, e.g., issue with video4linux2loopback streaming of a
IP camera in PTB message #18750.