Screen: Make video capture without explicit roi more robust.
If prerolling of the pipeline is disabled during 'OpenVideoCapture',
an explicit roi must be given, so 'GetCapturedImage' knows the size
of video frames.

Make this more user friendly: If size isn't known after opening
the video capture device due to skipped preroll or failed probe,
assign it during 'StartVideoCapture' on reception of the first
prerollbuffer, by probing the incoming buffers width x height.

-> Screen for 64-Bit Octave on Linux rebuilt and tested.