Windows: Test rebuilt most mex files, compile fixes.
Trivial compile fixes for Screen GStreamer stuff.
Change device-name to device-index for ksvideosrc
ksvideosrc is part of GStreamer 1.4.5, but still doesn’t work either
from command line or from within PTB with the iSight - from a Windows
VM that is. Might maybe work on a real Windows-7 box, but the
VirtualBox VM doesn’t cut it.